% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea_evaluationFunctions.R
\name{aggregateDEAeval}
\alias{aggregateDEAeval}
\title{aggregateDEAeval}
\usage{
aggregateDEAeval(res)
}
\arguments{
\item{res}{A list (per-dataset) of lists of DEA evaluation results.}
}
\value{
A list of aggregated results.
}
\description{
Aggregates DEA results (as produced by `evaluateDEA`) across datasets.
}
