% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregateResults}
\alias{aggregateResults}
\title{aggregateResults}
\usage{
aggregateResults(path = NULL, resfiles = NULL, pipDef = NULL)
}
\arguments{
\item{path}{Path to the folder containing the output of runPipeline}

\item{resfiles}{Path to *.stepIntermediateReturnObjects.rds files; the files 
are assumed to come from the `pipDef` pipeline, using the same alternatives
for all datasets.}

\item{pipDef}{A `pipelineDefinition` containing the aggregation methods.}
}
\value{
A list with a slot for each step for which there is an aggregation 
method, or (if no aggregation method available) a list of the 
`stepIntermediateReturnObjects` of `runPipeline`
}
\description{
Aggregates the `stepIntermediateReturnObjects` (and running times) of a 
`runPipeline` results. Results should be indicated either as a `path`` prefix
(with elapsed and pipeline info files along the results) or as a vector of 
paths to `stepIntermediateReturnObjects` files (`resfiles`).
}
