% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{buildCombMatrix}
\alias{buildCombMatrix}
\title{buildCombMatrix}
\usage{
buildCombMatrix(alt, returnIndexMatrix = FALSE)
}
\arguments{
\item{alt}{A named list of alternative parameter values}

\item{returnIndexMatrix}{Logical; whether to return a matrix of indices, 
rather than a data.frame of factors.}
}
\value{
a matrix or data.frame
}
\description{
Builds a matrix of parameter combinations from a list of alternative values.
}
\examples{
buildCombMatrix(list(param1=LETTERS[1:3], param2=1:2))
}
