% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{checkPipelinePackages}
\alias{checkPipelinePackages}
\title{checkPipelinePackages}
\usage{
checkPipelinePackages(alternatives, pipDef = NULL)
}
\arguments{
\item{alternatives}{A named list of alternative parameter values}

\item{pipDef}{An object of class `PipelineDefinition`.}
}
\value{
Logical.
}
\description{
Checks whether the packages required by a pipeline and its alternative 
methods are available.
}
\examples{
checkPipelinePackages(list(argument1="mean"), scrna_seurat_pipeline())
}
