% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{clustMetricsCorr}
\alias{clustMetricsCorr}
\title{Correlations across clustering evaluation metrics}
\value{
a list.
}
\description{
A list of two matrices containing, respectively, the Pearson and Spearman
pairwise correlations between various clustering evalution metrics, computed
across a wide range of scRNAseq clustering analyses (see reference).
}
\references{
See https://doi.org/10.1101/2020.02.02.930578
}
