% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ctrlgenes}
\alias{ctrlgenes}
\title{Lists of control genes}
\value{
a list.
}
\description{
Lists of mouse and human control genes (mitochondrial, ribosomal, 
protein-coding), as ensembl gene ids or official symbols, for computing cell
 QC.
}
