% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea_evaluationFunctions.R
\name{dea_evalPlot_sig}
\alias{dea_evalPlot_sig}
\title{dea_evalPlot_sig}
\usage{
dea_evalPlot_sig(res, what = c("TPR", "FDR"), threshold = 0.05, ...)
}
\arguments{
\item{res}{Aggregated results of the DEA pipeline}

\item{what}{What to plot}

\item{threshold}{Significance threshold to plot (default 0.05)}

\item{...}{Any further argument to `evalHeatmap`}
}
\value{
A Heatmap
}
\description{
This is a wrapper around `evalHeatmap` to enable the choice of the 
significance threshold when plotting results of the DEA pipeline.
}
\examples{
data("exampleDEAresults", package="pipeComp")
dea_evalPlot_sig( exampleDEAresults, agg.by=c("sva.method","dea.method"), 
                  threshold=0.05, row_split = "sva.method" )
}
