% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea_pipeline.R
\name{dea_pipeline}
\alias{dea_pipeline}
\title{dea_pipeline}
\usage{
dea_pipeline()
}
\value{
A `PipelineDefinition` object to be used with `runPipeline`.
}
\description{
The `PipelineDefinition` for bulk RNAseq differential expression analysis 
(DEA).
}
\examples{
pip <- dea_pipeline()
pip
}
