% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{defaultStepAggregation}
\alias{defaultStepAggregation}
\title{defaultStepAggregation}
\usage{
defaultStepAggregation(x)
}
\arguments{
\item{x}{A list of results per dataset, each containing a list (1 element 
per combination of parameters) of evaluation metrics (coercible to vectors 
or matrix).}
}
\value{
A data.frame.
}
\description{
defaultStepAggregation
}
