% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalHeatmap.R
\name{evalHeatmap}
\alias{evalHeatmap}
\title{evalHeatmap}
\usage{
evalHeatmap(
  res,
  step = NULL,
  what,
  what2 = NULL,
  agg.by = NULL,
  agg.fn = mean,
  scale = TRUE,
  value_format = "\%.2f",
  reorder_rows = FALSE,
  show_heatmap_legend = FALSE,
  show_column_names = FALSE,
  col = viridisLite::inferno(100),
  font_factor = 1,
  row_split = NULL,
  shortNames = TRUE,
  value_cols = c("black", "white"),
  title = NULL,
  anno_legend = TRUE,
  ...
)
}
\arguments{
\item{res}{Aggregated pipeline results (i.e. the output of `runPipeline` or
`aggregateResults`)}

\item{step}{Name of the step for which to plot the evaluation results}

\item{what}{What metric to plot}

\item{what2}{If the step has more than one benchmark data.frame, which one
to use.}

\item{agg.by}{Aggregate results by these columns (default no aggregation)}

\item{agg.fn}{Function for aggregation (default mean)}

\item{scale}{Logical; whether to scale columns (default FALSE)}

\item{value_format}{Format for displaying cells' values (use 
`value_format=""` to disable)}

\item{reorder_rows}{Logical; whether to sort rows (default FALSE). The row 
names themselves can also be passed to specify an order, or a 
`ComplexHeatmap`.}

\item{show_heatmap_legend}{Passed to `Heatmap` (default FALSE)}

\item{show_column_names}{Passed to `Heatmap` (default FALSE)}

\item{col}{Colors for the heatmap}

\item{font_factor}{A scaling factor applied to fontsizes (default 1)}

\item{row_split}{Optional column (included in `agg.by`) by which to split
the rows.}

\item{shortNames}{Logical; whether to use short row names (with only
the parameter values instead of the parameter name and value pairs), default
TRUE.}

\item{value_cols}{A vector of length 2 indicating the colors of the values
(above and below the mean), if printed}

\item{title}{Plot title}

\item{anno_legend}{Logical; whether to plot the legend for the datasets}

\item{...}{Passed to `Heatmap`}
}
\value{
A Heatmap
}
\description{
General heatmap representation of aggregated evaluation results.
}
\examples{
data("exampleDEAresults", package="pipeComp")
evalHeatmap( exampleDEAresults, what=c("logFC.pearson", "TPR","FDR"), 
             agg.by=c("sva.method","dea.method"), row_split = "sva.method" )
}
