% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna_evaluationFunctions.R
\name{evaluateClustering}
\alias{evaluateClustering}
\title{evaluateClustering}
\usage{
evaluateClustering(x, tl)
}
\arguments{
\item{x}{The clustering labels}

\item{tl}{The true labels}
}
\value{
A numeric vector.
}
\description{
Evaluates a clustering using 'true' labels
}
