% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna_utils.R
\name{getDimensionality}
\alias{getDimensionality}
\title{getDimensionality}
\usage{
getDimensionality(dat, method, maxDims = NULL)
}
\arguments{
\item{dat}{A Seurat or SCE object with a pca embedding.}

\item{method}{The dimensionality method to use.}

\item{maxDims}{Deprecated and ignored.}
}
\value{
An integer.
}
\description{
Returns the estimated intrinsic dimensionality of a dataset.
}
