% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getQualitativePalette}
\alias{getQualitativePalette}
\title{getQualitativePalette}
\usage{
getQualitativePalette(nbcolors)
}
\arguments{
\item{nbcolors}{A positive integer indicating the number of colors}
}
\value{
A vector of colors
}
\description{
Returns a qualitative color palette of the given size. If less than 23 colors
 are required, the colors are based on Paul Tol's palettes. If more, the 
 `randomcoloR` package is used.
}
\examples{
getQualitativePalette(5)
}
