% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalknn.R
\name{knnPrecision}
\alias{knnPrecision}
\title{knnPrecision}
\usage{
knnPrecision(pred, truth, kmax = NULL, isKNN = FALSE)
}
\arguments{
\item{pred, truth}{The matrices of predicted and true values, respectively, 
with variables as columns. Alternatively, if `isKNN=TRUE`, matrices of 
nearest neighbor indices (the first column being the first neighbor or each 
item), as produced by \code{\link[BiocNeighbors]{findKNN}$index}. Either 
way, the rows are assumed to correspond between the two objects.}

\item{kmax}{The maximum k nearest neighbor to consider.}

\item{isKNN}{Logical; whether `pred` and `truth` are already computed kNNs.
If FALSE (default), both kNN are first computed.}
}
\value{
A list containing 1) a matrix of kNN precisions (as %) for each 
item, at each value of k, and 2) the area under the PRK curve for each item.
}
\description{
Calculates the precision-at-k (PRK) curve, i.e. the proportion of correct k 
nearest neighbors across values of k, as a measure of the accuracy of a 
kNN network judged against a 'true' one.
}
\examples{
x <- runif(500)*10
e1 <- matrix(rpois(length(x),x),ncol=10)
e2 <- matrix(rpois(length(x),x),ncol=10)
ac <- knnPrecision(e1,e2)
hist(ac$AUC, xlab="Area under the PRK curve")
boxplot(as.data.frame(ac$precisions), xlab="Number of nearest neighbors", 
        ylab="Proportion correct", outline=FALSE)
}
