% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{mergePipelineResults}
\alias{mergePipelineResults}
\title{mergePipelineResults}
\usage{
mergePipelineResults(res1, res2)
}
\arguments{
\item{res1}{A list of pipeline results, as produced by `readPipelineResults`}

\item{res2}{A list of pipeline results, as produced by `readPipelineResults`}
}
\value{
A list of pipeline results.
}
\description{
mergePipelineResults
}
