% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipelineDefinition.R
\name{mockPipeline}
\alias{mockPipeline}
\title{mockPipeline}
\usage{
mockPipeline()
}
\value{
a `PipelineDefinition`
}
\description{
A mock `PipelineDefinition` for use in examples.
}
\examples{
mockPipeline()
}
