% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{parsePipNames}
\alias{parsePipNames}
\title{parsePipNames}
\usage{
parsePipNames(x, setRowNames = FALSE, addcolumns = NULL)
}
\arguments{
\item{x}{The names to parse, or a data.frame with the names to parse as 
row.names. All names are expected to contain the same parameters.}

\item{setRowNames}{Logical; whether to set original names as row.names of
the output data.frame (default FALSE)}

\item{addcolumns}{An optional data.frame of `length(x)` rows to cbind to the
output.}
}
\value{
A data.frame
}
\description{
Parses the names of analyses performed through `runPipeline` to extract a 
data.frame of parameter values (with decent classes).
}
\examples{
my_names <- c("param1=A;param2=5","param1=B;param2=0")
parsePipNames(my_names)
}
