% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalHeatmap.R
\name{renameHrows}
\alias{renameHrows}
\title{renameHrows}
\usage{
renameHrows(h, f = identity)
}
\arguments{
\item{h}{A `Heatmap` or `HeatmapList`}

\item{f}{The function to apply on row labels.}
}
\value{
the modified `h`
}
\description{
Applies a function to the row labels of a Heatmap or HeatmapList
}
\examples{
data("exampleResults", package="pipeComp")
H <- evalHeatmap(exampleResults, what=c("ARI", "MI"), agg.by="norm")
H
H <- renameHrows(H, function(x) gsub("^norm\\\\.", "", x))
H
}
