% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna_plot.R
\name{scrna_evalPlot_DR}
\alias{scrna_evalPlot_DR}
\title{scrna_evalPlot_DR}
\usage{
scrna_evalPlot_DR(res, what = c("auto", "varExpl", "minSilWidth",
  "meanSilWidth", "medianSilWidth", "maxSilWidth", "log10_total_counts",
  "log10_total_features", "total_features"), reorder_rows = TRUE,
  reorder_columns = NULL, agg.by = NULL, agg.fn = mean,
  scale = FALSE, show_heatmap_legend = FALSE, value_format = "\%.2f",
  col = NULL, col_title_fontsize = 11, value_cols = c("black",
  "white"), anno_legend = TRUE, ...)
}
\arguments{
\item{res}{Aggregated pipeline results (i.e. the output of `runPipeline` or
`aggregateResults`)}

\item{what}{What to plot (default plots main metrics)}

\item{reorder_rows}{Logical; whether to sort rows (default TRUE)}

\item{reorder_columns}{Logical; whether to sort columns}

\item{agg.by}{Aggregate results by these columns (default no aggregation)}

\item{agg.fn}{Function for aggregation (default mean)}

\item{scale}{Logical; whether to scale columns (default FALSE)}

\item{show_heatmap_legend}{Passed to `Heatmap`}

\item{value_format}{Format for displaying cells' values (use 
`value_format=""` to disable)}

\item{col}{Colors for the heatmap}

\item{col_title_fontsize}{Fontsite of the column titles}

\item{value_cols}{A vector of length 2 indicating the colors of the values
(above and below the mean), if printed}

\item{...}{Passed to `Heatmap`}

\item{covar}{Covariate of interest (used only if `what` is  'covar' or 
'covarRes'}
}
\value{
One or several `Heatmap` object.
}
\description{
Plotting aggregated evaluation results at the level of dimensionality 
reduction for the scRNA pipelines.
}
