% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna_seurat_pipeline.R
\name{scrna_seurat_pipeline}
\alias{scrna_seurat_pipeline}
\title{scrna_seurat_pipeline}
\usage{
scrna_seurat_pipeline(saveDimRed = FALSE)
}
\arguments{
\item{saveDimRed}{Logical; whether to save the dimensionality reduction for 
each analysis (default FALSE)}
}
\description{
The default SC clustering pipeline, with steps for filtering, normalization, feature selection,
dimensionality reduction, and clustering.
A pipeline definition is a named list of functions in the order in which they should be run. The
first argument of each function should be `x`, which refers to the output of the previous function
(or the initial object for the first step). Any other arguments will be arguments 
that can be varied in the pipeline comparison: they can be character, numeric or logical vectors of
length 1 (e.g. the function name for a method, the number of dimensions, etc).
The default pipeline has the following steps and arguments:
* doublet: `doubletmethod` (name of the doublet removal function)
* filtering: `filt` (name of the filtering function, or filter string)
* normalization: `norm` (name of the normalization function)
* selection: `sel` (name of the selection function, or variable of rowData on which to select) and 
 `selnb` (number of features to select)
* dimreduction: `dr` (name of the dimensionality reduction function) and `maxdim` (maximum number 
of components to compute)
* clustering: `clustmethod` (name of the clustering function), `dims` (number of dimensions to use),
 `k` (number of nearest neighbors to use, if applicable), `steps` (number of steps in the random 
 walk, if applicable), `resolution` (resolution, if applicable), `min.size` (minimum cluster size,
 if applicable)
}
