% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna_plot.R
\name{describeDatasets}
\alias{describeDatasets}
\title{describeDatasets}
\usage{
describeDatasets(sces, pt.size = 0.3, ...)
}
\arguments{
\item{sces}{A character vector of paths to SCE rds files, or a list of SCEs}

\item{pt.size}{Point size (for reduced dims)}

\item{...}{Passed to geom_point()}
}
\value{
A plot_grid output
}
\description{
Plots descriptive information about the datasets
}
