% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea_evaluationFunctions.R
\name{evaluateDEA}
\alias{evaluateDEA}
\title{evaluateDEA}
\usage{
evaluateDEA(dea, truth = NULL, th = c(0.01, 0.05, 0.1))
}
\arguments{
\item{dea}{Expects a data.frame with logFC and FDR, as produced by `edgeR::topTags`, `limma::topTable` or 
`DESeq2::results`.}

\item{truth}{A data.frame containing the columns `expected.beta` (real logFC)
and `isDE` (logical indicating whether there is a difference or not; accepts
NA values)}

\item{th}{The significance thresholds for which to compute the metrics.}
}
\value{
A list with two slots: `logFC` (vector of metrics on logFC) and 
`significance` table of significance-related statistics.
}
\description{
Evaluates a differential expression analysis (DEA).
}
