% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{mergePipelineResults}
\alias{mergePipelineResults}
\title{mergePipelineResults}
\usage{
mergePipelineResults(res1, res2)
}
\arguments{
\item{res1}{A list of pipeline results, as produced by `readPipelineResults`}

\item{res2}{A list of pipeline results, as produced by `readPipelineResults`}
}
\value{
A list of pipeline results.
}
\description{
mergePipelineResults
}
\examples{
# we produce 2 mock pipeline results:
pip <- mockPipeline()
datasets <- list( ds1=1:3, ds2=c(5,10,15) )
tmpdir1 <- paste0(tempdir(),'/')
res <- runPipeline(datasets, pipelineDef=pip, output.prefix=tmpdir1,
                   alternatives=list() )
alternatives <- list(meth1=c('log2','sqrt'), meth2='cumsum')
tmpdir2 <- paste0(tempdir(),'/')
res <- runPipeline(datasets, alternatives, pip, output.prefix=tmpdir2)
# we read the evaluation files:
res1 <- readPipelineResults(tmpdir1)
res2 <- readPipelineResults(tmpdir2)
# we merge them:
res <- mergePipelineResults(res1,res2)
# and we aggregate:
res <- aggregatePipelineResults(res)
}
