% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPipeline.R
\name{pipeComp-package}
\alias{pipeComp-package}
\alias{pipeComp}
\title{pipeComp - a framework for pipeline benchmarking}
\description{
\pkg{pipeComp} is a simple framework to facilitate the comparison of 
pipelines involving various steps and parameters. It was initially developed 
to benchmark single-cell RNA sequencing pipelines, and contains pre-defined
\code{\link{PipelineDefinition}}s and functions to that effect, but could be
applied to any context. See `vignette("pipeComp")` for an introduction.
}
\author{
Pierre-Luc Germain \email{pierre-luc.germain@hest.ethz.ch}

Anthony Sonrel \email{anthony.sonrel@uzh.ch}

Mark D. Robinson \email{mark.robinson@imls.uzh.ch}
}
