% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipelineDefinition.R
\name{PipelineDefinition-methods}
\alias{PipelineDefinition-methods}
\alias{PipelineDefinition-method}
\alias{show,PipelineDefinition-method}
\alias{names,PipelineDefinition-method}
\alias{names<-,PipelineDefinition,ANY-method}
\alias{$,PipelineDefinition-method}
\alias{length,PipelineDefinition-method}
\alias{[,PipelineDefinition,ANY,ANY,ANY-method}
\alias{as.list,PipelineDefinition-method}
\alias{arguments}
\alias{arguments,PipelineDefinition-method}
\alias{defaultArguments}
\alias{defaultArguments<-}
\alias{defaultArguments,PipelineDefinition-method}
\alias{defaultArguments<-,PipelineDefinition-method}
\alias{stepFn}
\alias{stepFn,PipelineDefinition-method}
\alias{stepFn<-}
\alias{stepFn<-,PipelineDefinition-method}
\title{Methods for \code{\link{PipelineDefinition}} class}
\usage{
\S4method{show}{PipelineDefinition}(object)

\S4method{names}{PipelineDefinition}(x)

\S4method{names}{PipelineDefinition,ANY}(x) <- value

\S4method{$}{PipelineDefinition}(x, name)

\S4method{length}{PipelineDefinition}(x)

\S4method{[}{PipelineDefinition,ANY,ANY,ANY}(x, i)

\S4method{as.list}{PipelineDefinition}(x)

arguments(object)

\S4method{arguments}{PipelineDefinition}(object)

defaultArguments(object)

defaultArguments(object) <- value

\S4method{defaultArguments}{PipelineDefinition}(object)

\S4method{defaultArguments}{PipelineDefinition}(object) <- value

stepFn(object, step, type)

\S4method{stepFn}{PipelineDefinition}(object, step, type)

stepFn(object, step, type) <- value

\S4method{stepFn}{PipelineDefinition}(object, step, type) <- value
}
\arguments{
\item{object}{An object of class \code{\link{PipelineDefinition}}}

\item{x}{An object of class \code{\link{PipelineDefinition}}}

\item{value}{Replacement values}

\item{name}{The step name}

\item{i}{The index(es) of the steps}

\item{step}{The name of the step for which to set or get the function}

\item{type}{The type of function to set/get, either `functions`, 
`evaluation`, `aggregation`, `descriptions`, or `initiation` (will parse 
partial matches)}
}
\value{
Depends on the method.
}
\description{
Methods for \code{\link{PipelineDefinition}} class
}
\examples{
pd <- mockPipeline()
length(pd)
names(pd)
pd$step1
pd[2:1]
}
\seealso{
\code{\link{PipelineDefinition}}, \code{\link{addPipelineStep}}
}
