% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna_plot.R
\name{scrna_evalPlot_clust}
\alias{scrna_evalPlot_clust}
\title{scrna_evalPlot_clust}
\usage{
scrna_evalPlot_clust(res, what = "auto", agg.by = NULL,
  agg.fn = mean, scale = FALSE, value_format = "\%.2f",
  reorder_rows = TRUE, show_heatmap_legend = FALSE,
  col = viridisLite::inferno(100), col_title_fontsize = 12, ...)
}
\arguments{
\item{res}{Aggregated pipeline results (i.e. the output of `runPipeline` or
`aggregateResults`)}

\item{what}{What to plot (default plots main metrics)}

\item{agg.by}{Aggregate results by these columns (default no aggregation)}

\item{agg.fn}{Function for aggregation (default mean)}

\item{scale}{Logical; whether to scale columns (default FALSE)}

\item{value_format}{Format for displaying cells' values (use 
`value_format=""` to disable)}

\item{reorder_rows}{Logical; whether to sort rows (default TRUE)}

\item{show_heatmap_legend}{Passed to `Heatmap`}

\item{col}{Colors for the heatmap}

\item{col_title_fontsize}{Fontsize of column titles.}

\item{...}{Passed to `Heatmap`}
}
\value{
One or several `Heatmap` object.
}
\description{
Plotting aggregated evaluation results at the level of clustering for the 
scRNA pipelines.
}
