% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna_plot.R
\name{scrna_evalPlot_filtering}
\alias{scrna_evalPlot_filtering}
\title{scrna_evalPlot_filtering}
\usage{
scrna_evalPlot_filtering(res, steps = c("doublet", "filtering"),
  clustMetric = "mean_F1", returnTable = FALSE)
}
\arguments{
\item{res}{Aggregated pipeline results (i.e. the output of `runPipeline` or
`aggregateResults`)}

\item{steps}{Steps to include (default 'doublet' and 'filtering'); other 
steps will be averaged.}

\item{clustMetric}{Clustering accuracy metric to use (default `mean_F1``)}

\item{returnTable}{Logical; whether to return the data.frame rather than plot.}
}
\value{
A ggplot, or a data.frame if `returnTable=TRUE`
}
\description{
scrna_evalPlot_filtering
}
